# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "API description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/providers": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "list of providers",
                "description": "Returns the entire list of providers",
                "operationId": "Provider.all",
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Providers"
                ],
                "summary": "Create a new provider",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "Provider.create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url",
                                    "domain",
                                    "logoutUrl",
                                    "secret_key"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the provider",
                                        "type": "string",
                                        "example": "Example"
                                    },
                                    "url": {
                                        "description": "URL of the provider",
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "domain": {
                                        "description": "Domain of the provider",
                                        "type": "string",
                                        "example": "example.com"
                                    },
                                    "protocol": {
                                        "description": "Protocol of the provider",
                                        "type": "string",
                                        "example": "http"
                                    },
                                    "logoutUrl": {
                                        "description": "Logout URL of the provider",
                                        "type": "string",
                                        "example": "https://example.com/logout"
                                    },
                                    "secret_key": {
                                        "description": "Signature key of the provider",
                                        "type": "string",
                                        "example": "2d6f5d6f8d6f2d6f5d6f8d6f2d6f5d6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/providers/{id}": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Get provider by id",
                "description": "Returns provider details by id",
                "operationId": "Provider.find",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Providers"
                ],
                "summary": "Update provider by id",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "Provider.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url",
                                    "domain",
                                    "logoutUrl"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Provider name",
                                        "type": "string",
                                        "example": "Portale Newtimegroup"
                                    },
                                    "url": {
                                        "description": "Provider url",
                                        "type": "string",
                                        "example": "https://portale.newtimegroup.it"
                                    },
                                    "domain": {
                                        "description": "Provider domain",
                                        "type": "string",
                                        "example": "portale.newtimegroup.it"
                                    },
                                    "protocol": {
                                        "description": "Provider protocol",
                                        "type": "string",
                                        "example": "http"
                                    },
                                    "logoutUrl": {
                                        "description": "URL for logout",
                                        "type": "string",
                                        "example": "https://portale.newtimegroup.it/logout/"
                                    },
                                    "secret_key": {
                                        "description": "Signature key for JWT token. Must be 32 characters long. Leave empty to not change it.",
                                        "type": "string",
                                        "example": "2d6f5d6f8d6f2d6f5d6f8d6f2d6f5d6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Providers"
                ],
                "summary": "Delete provider by id",
                "description": "__*Security:*__ __*can be used only by clients with 'admin' role*__",
                "operationId": "Provider.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Operation successful",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 Client Credentials Flow",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}